+++ src/main/java/edu/unh/cs/treccar/proj/cluster/ClusterResult.java
@@ -1,4 +1,4 @@
-package edu.unh.cs.treccar.proj;
+package edu.unh.cs.treccar.proj.cluster;
 
 import java.util.ArrayList;
