+++ src/main/java/edu/unh/cs/treccar/proj/similarities/JiangConrathSimilarity.java
@@ -1,8 +1,10 @@
-package edu.unh.cs.treccar.proj;
+package edu.unh.cs.treccar.proj.similarities;
 
 import java.util.ArrayList;
 
 import edu.unh.cs.treccar.Data;
+import edu.unh.cs.treccar.proj.util.ParaPair;
+import edu.unh.cs.treccar.proj.util.ParaUtilities;
 
 /**
  * This class implements the JiangConrath Semantic Similarity measure
