+++ src/main/java/edu/unh/cs/treccar/proj/similarities/LinSimilarity.java
@@ -1,16 +1,18 @@
-package edu.unh.cs.treccar.proj;
+package edu.unh.cs.treccar.proj.similarities;
 
 import java.util.ArrayList;
 
 import edu.unh.cs.treccar.Data;
+import edu.unh.cs.treccar.proj.util.ParaPair;
+import edu.unh.cs.treccar.proj.util.ParaUtilities;
 
 /**
- * This class implements the LeacockChodorow Semantic Similarity measure 
+ * This class implements the Lin Semantic Similarity measure 
  * and returns the measure between two paragraphs.
  * @author Shubham Chatterjee
  *
  */
-public class LeacockChodorowSimilarity implements SimilarityFunction
+public class LinSimilarity implements SimilarityFunction
 {
 	private static ArrayList<String> paraText1;
 	private static ArrayList<String> paraText2;
@@ -20,7 +22,7 @@ public class LeacockChodorowSimilarity implements SimilarityFunction
 	/**
 	 * @param pp ParaPair A ParaPair object representing a pair of paragraphs
 	 * @param list ArrayList<Data.Paragraph> Conatins a list of Data.Paragraph objects 
-	 * @return LeacockChodorow score between two paragraphs
+	 * @return Lin score between two paragraphs
 	 */
 	
 	public double simScore(ParaPair pp, ArrayList<Data.Paragraph> list)
@@ -43,10 +45,9 @@ public class LeacockChodorowSimilarity implements SimilarityFunction
 	private static double getParaScore(ArrayList<String> list1, ArrayList<String> list2)
 	{
 		double s = 0.0d;
-		scores = new ArrayList<Double>();
 		for(String w1 : list1)
 			for(String w2 : list2)
-				scores.add(WordSemanticSimilarity.findSimilarity(w1, w2, "lc"));
+				scores.add(WordSemanticSimilarity.findSimilarity(w1, w2, "lin"));
 		s = findMeanScore(scores);
 		return s;
 	}
@@ -67,3 +68,4 @@ public class LeacockChodorowSimilarity implements SimilarityFunction
 		return mean;
 	}
 }
+
