+++ src/main/java/edu/unh/cs/treccar/proj/similarities/PathSimilarity.java
@@ -1,16 +1,19 @@
-package edu.unh.cs.treccar.proj;
+package edu.unh.cs.treccar.proj.similarities;
 
 import java.util.ArrayList;
+
 import edu.unh.cs.treccar.Data;
+import edu.unh.cs.treccar.proj.util.ParaPair;
+import edu.unh.cs.treccar.proj.util.ParaUtilities;
 
 /**
- * This class implements the JiangConrath Semantic Similarity measure 
+ * This class implements the Path Semantic Similarity measure 
  * and returns the measure between two paragraphs.
  * @author Shubham Chatterjee
  *
  */
 
-public class WuPalmerSimilarity implements SimilarityFunction
+public class PathSimilarity implements SimilarityFunction
 {
 	private static ArrayList<String> paraText1;
 	private static ArrayList<String> paraText2;
@@ -20,8 +23,9 @@ public class WuPalmerSimilarity implements SimilarityFunction
 	/**
 	 * @param pp ParaPair A ParaPair object representing a pair of paragraphs
 	 * @param list ArrayList<Data.Paragraph> Conatins a list of Data.Paragraph objects 
-	 * @return WuPalmer score between two paragraphs
+	 * @return Path score between two paragraphs
 	 */
+	
 	public double simScore(ParaPair pp, ArrayList<Data.Paragraph> list)
 	{
 		paraText1 = ParaUtilities.getParaText1(pp, list);
@@ -31,6 +35,7 @@ public class WuPalmerSimilarity implements SimilarityFunction
 		
 		return score;
 	}
+	
 	/**
 	 * 
 	 * @param list1 ArrayList<String> List of words from first paragraph
@@ -43,10 +48,11 @@ public class WuPalmerSimilarity implements SimilarityFunction
 		double s = 0.0d;
 		for(String w1 : list1)
 			for(String w2 : list2)
-				scores.add(WordSemanticSimilarity.findSimilarity(w1, w2, "wp"));
+				scores.add(WordSemanticSimilarity.findSimilarity(w1, w2, "path"));
 		s = findMeanScore(scores);
 		return s;
 	}
+	
 	/**
 	 * 
 	 * @param list ArrayList<Double> List of scores between words of two paragraphs
