+++ src/main/java/edu/unh/cs/treccar/proj/similarities/ResnikSimilarity.java
@@ -1,15 +1,19 @@
-package edu.unh.cs.treccar.proj;
+package edu.unh.cs.treccar.proj.similarities;
 
 import java.util.ArrayList;
 
 import edu.unh.cs.treccar.Data;
+import edu.unh.cs.treccar.proj.util.ParaPair;
+import edu.unh.cs.treccar.proj.util.ParaUtilities;
+
 /**
- * This class implements the Lesk Semantic Similarity measure 
+ * This class implements the Resnik Semantic Similarity measure 
  * and returns the measure between two paragraphs.
  * @author Shubham Chatterjee
  *
  */
-public class LeskSimilarity implements SimilarityFunction
+
+public class ResnikSimilarity implements SimilarityFunction
 {
 	private static ArrayList<String> paraText1;
 	private static ArrayList<String> paraText2;
@@ -19,7 +23,7 @@ public class LeskSimilarity implements SimilarityFunction
 	/**
 	 * @param pp ParaPair A ParaPair object representing a pair of paragraphs
 	 * @param list ArrayList<Data.Paragraph> Conatins a list of Data.Paragraph objects 
-	 * @return Lesk score between two paragraphs
+	 * @return Resnik score between two paragraphs
 	 */
 	
 	public double simScore(ParaPair pp, ArrayList<Data.Paragraph> list)
@@ -31,7 +35,6 @@ public class LeskSimilarity implements SimilarityFunction
 		
 		return score;
 	}
-	
 	/**
 	 * 
 	 * @param list1 ArrayList<String> List of words from first paragraph
@@ -42,10 +45,9 @@ public class LeskSimilarity implements SimilarityFunction
 	private static double getParaScore(ArrayList<String> list1, ArrayList<String> list2)
 	{
 		double s = 0.0d;
-		scores = new ArrayList<Double>();
 		for(String w1 : list1)
 			for(String w2 : list2)
-				scores.add(WordSemanticSimilarity.findSimilarity(w1, w2, "lesk"));
+				scores.add(WordSemanticSimilarity.findSimilarity(w1, w2, "res"));
 		s = findMeanScore(scores);
 		return s;
 	}
