+++ src/main/java/edu/unh/cs/treccar/proj/similarities/SimilarityFunction.java
@@ -1,8 +1,9 @@
-package edu.unh.cs.treccar.proj;
+package edu.unh.cs.treccar.proj.similarities;
 
 import java.util.ArrayList;
 
 import edu.unh.cs.treccar.Data;
+import edu.unh.cs.treccar.proj.util.ParaPair;
 
 public interface SimilarityFunction {
 	public double simScore(ParaPair pp, ArrayList<Data.Paragraph> list);
