+++ src/main/java/edu/unh/cs/treccar/proj/similarities/Word2VecSim.java
@@ -1,8 +1,9 @@
-package edu.unh.cs.treccar.proj;
+package edu.unh.cs.treccar.proj.similarities;
 
 import java.util.ArrayList;
 
 import edu.unh.cs.treccar.Data.Paragraph;
+import edu.unh.cs.treccar.proj.util.ParaPair;
 
 public class Word2VecSim implements SimilarityFunction {
